add margin to obstacle check
