+++ lib/check-each-pcb-trace-non-overlapping/are-bounds-overlapping.ts
@@ -1,10 +1,19 @@
 import type { Bounds } from "lib/data-structures/SpatialIndex"
 
-export const areBoundsOverlapping = (bounds1: Bounds, bounds2: Bounds) => {
+/**
+ * Are two bounds overlapping?
+ *
+ * If there is a margin, we increase one of the bounds by the margin
+ */
+export const areBoundsOverlapping = (
+  bounds1: Bounds,
+  bounds2: Bounds,
+  margin = 0,
+) => {
   return (
-    bounds1.minX <= bounds2.maxX &&
-    bounds1.maxX >= bounds2.minX &&
-    bounds1.minY <= bounds2.maxY &&
-    bounds1.maxY >= bounds2.minY
+    bounds1.minX <= bounds2.maxX + margin &&
+    bounds1.maxX >= bounds2.minX - margin &&
+    bounds1.minY <= bounds2.maxY + margin &&
+    bounds1.maxY >= bounds2.minY - margin
   )
 }
